Compilation error fixed for 64 bit architectures.

+++ src/plugins/simulator/physics_engines/dynamics2d/dynamics2d_engine.cpp
@@ -215,7 +215,7 @@ namespace argos {
    /****************************************/
    /****************************************/
 
-   UInt32 CDynamics2DEngine::GetNumPhysicsModels() {
+   size_t CDynamics2DEngine::GetNumPhysicsModels() {
       return m_tPhysicsModels.size();
    }
